# java-lamp-api/public/swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: JavaLamp API
  description: >
    #### Backend for JavaLamp
basePath: /
parameters:
  id:
    name: id
    in: path
    description: id
    type: integer
    required: true
definitions:
  Links:
    type: object
  UsersCollection:
    type: object
    properties:
      links:
        $ref: '#/definitions/Links'
      data:
        type: array
        items:
          $ref: '#/definitions/UsersResource'
  UserDocument:
    type: object
    properties:
      links:
        $ref: '#/definitions/Links'
      data:
        $ref: '#/definitions/UsersResource'
  UsersResource:
    type: object
    properties:
      type:
        type: string
        enum:
        - user
      id:
        type: integer
      attributes:
        $ref: '#/definitions/User'
      links:
        $ref: '#/definitions/Links'
  User:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
      token:
        type: string
  Error:
    type: object
    properties:
      error:
        type: array
        items:
          type: object
          properties:
            status:
              type: string
            source:
              type: object
              properties:
                pointer:
                  type: string
            title:
              type: string
            details:
              type: string
paths:
  '/users':
    get:
      description: >
        Get all users
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UsersCollection'
    post:
      description: >
        Posts a new user
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserDocument'
        '400':
          description: Validation error
          schema:
            $ref: '#/definitions/Error'
      parameters:
        - name: Payload
          in: body
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  type:
                    type: string
                    enum:
                    - user
                  attributes:
                    type: object
                    properties:
                      name:
                        type: string
                      email:
                        type: string
                      password:
                        type: string
                      token:
                        type: string
  '/users/{id}':
    get:
      parameters:
        - $ref: '#/parameters/id'
      description: >
        Get one user
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UsersCollection'
    put:
      parameters:
        - $ref: '#/parameters/id'
        - name: Payload
          in: body
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  type:
                    type: string
                    enum:
                    - user
                  attributes:
                    type: object
                    properties:
                      name:
                        type: string
                      email:
                        type: string
                      password:
                        type: string
                      token:
                        type: string
    delete:
      parameters:
        - $ref: '#/parameters/id'
      description: >
        Delete one user
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UsersCollection'
  '/questions':
    get:
      description: >
        Get all questions
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/QuestionsCollection'
    post:
      description: >
        Posts a new question
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/QuestionDocument'
        '400':
          description: Validation error
          schema:
            $ref: '#/definitions/Error'
      parameters:
        - name: Payload
          in: body
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  type:
                    type: string
                    enum:
                    - question
                  attributes:
                    type: object
                    properties:
                      title:
                        type: string
                      questionText:
                        type: string
                      sampleCode:
                        type: string
                      test:
                        type: string
  '/questions/{id}':
    get:
      parameters:
        - $ref: '#/parameters/id'
      description: >
        Get one question
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/QuestionsCollection'
    put:
      parameters:
        - $ref: '#/parameters/id'
        - name: Payload
          in: body
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  type:
                    type: string
                    enum:
                    - question
                  attributes:
                    type: object
                    properties:
                      title:
                        type: string
                      questionText:
                        type: string
                      sampleCode:
                        type: string
                      test:
                        type: string
    delete:
      parameters:
        - $ref: '#/parameters/id'
      description: >
        Delete one question
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/QuestionsCollection'
  '/answers':
    get:
      description: >
        Get all answers
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/AnswersCollection'
    post:
      description: >
        Posts a new answer
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/AnswerDocument'
        '400':
          description: Validation error
          schema:
            $ref: '#/definitions/Error'
      parameters:
        - name: Payload
          in: body
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  type:
                    type: string
                    enum:
                    - answer
                  attributes:
                    type: object
                    properties:
                      answerCode:
                        type: string
                      answerText:
                        type: string
  '/answers/{id}':
    get:
      parameters:
        - $ref: '#/parameters/id'
      description: >
        Get one answer
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/AnswersCollection'
    put:
      parameters:
        - $ref: '#/parameters/id'
        - name: Payload
          in: body
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  type:
                    type: string
                    enum:
                    - answer
                  attributes:
                    type: object
                    properties:
                      answerCode:
                        type: string
                      answerText:
                        type: string
    delete:
      parameters:
        - $ref: '#/parameters/id'
      description: >
        Delete one answer
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/AnswersCollection'